# clinic-api/docs/swagger.yaml
basePath: /api
definitions:
...
  dentist.CreateInput:
    properties:
      bio:
        example: Especialista em ortodontia
        type: string
      email:
        example: ana.souza@clinicasorriso.com.br
        type: string
      name:
        example: Dra. Ana Souza
        type: string
      phone:
        example: (11) 98888-0001
        type: string
      specialties:
        example:
        - Ortodontia
        - Invisalign
        items:
          type: string
        type: array
      years_of_experience:
        example: 12
        type: integer
    type: object
  dentist.RolesInput:
    properties:
      is_administrator:
        example: true
        type: boolean
      is_legal_representative:
        example: true
        type: boolean
    type: object
  dentist.UpdateInput:
    properties:
      bio:
        example: Especialista em ortodontia
        type: string
      email:
        example: ana.souza@clinicasorriso.com.br
        type: string
      name:
        example: Dra. Ana Souza
        type: string
      phone:
        example: (11) 98888-0001
        type: string
      specialties:
        items:
          type: string
        type: array
      years_of_experience:
        example: 12
        type: integer
    type: object
  dentist.dentistResponse:
    properties:
      bio:
        example: Especialista em ortodontia
        type: string
      clinic_id:
        example: a0000000-0000-0000-0000-000000000001
        type: string
      created_at:
        type: string
      email:
        example: ana.souza@clinicasorriso.com.br
        type: string
      id:
        example: a0000000-0000-0000-0000-000000000011
        type: string
      is_administrator:
        example: true
        type: boolean
      is_legal_representative:
        example: true
        type: boolean
      name:
        example: Dra. Ana Souza
        type: string
      phone:
        example: (11) 98888-0001
        type: string
      specialties:
        items:
          type: string
        type: array
      updated_at:
        type: string
      years_of_experience:
        example: 12
        type: integer
    type: object
  dentist.listResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dentist.dentistResponse'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  problem.Details:
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      detail:
        example: request validation failed
        type: string
      errors:
        items:
          $ref: '#/definitions/problem.FieldError'
        type: array
      status:
        example: 400
        type: integer
      title:
        example: Bad Request
        type: string
      type:
        example: about:blank
        type: string
    type: object
  problem.FieldError:
    properties:
      detail:
        example: is required
        type: string
      field:
        example: clinic_id
        type: string
    type: object
info:
  contact: {}
  description: |-
    API de gestão de clínica odontológica: clínicas, dentistas e pagamentos via Pix (simulado).
    Todos os endpoints de negócio ficam sob o prefixo /api; o único fora dele é o health check em /health.
    Erros seguem o formato RFC 9457 (application/problem+json) em todo o corpo, com um campo "code" estável para tratamento programático além do "status" HTTP.
  title: Clinic API
  version: "1.0"
paths:
...
  /clinics/{clinic_id}/dentists:
    get:
      parameters:
      - description: Clinic ID
        format: uuid
        in: path
        name: clinic_id
        required: true
        type: string
      - default: 20
        description: Page size
        in: query
        maximum: 100
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      - description: Filter by administrator flag
        in: query
        name: is_administrator
        type: boolean
      - description: Filter by legal representative flag
        in: query
        name: is_legal_representative
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dentist.listResponse'
        "400":
          description: clinic_id is not a valid UUID
          schema:
            $ref: '#/definitions/problem.Details'
        "404":
          description: clinic not found
          schema:
            $ref: '#/definitions/problem.Details'
      summary: List dentists of a clinic (paginated)
      tags:
      - dentists
    post:
      consumes:
      - application/json
      description: Creates a dentist attached to the given clinic. Set is_administrator/is_legal_representative
        afterwards via the roles endpoint if needed.
      parameters:
      - description: Clinic ID
        format: uuid
        in: path
        name: clinic_id
        required: true
        type: string
      - description: Dentist to create
        in: body
        name: dentist
        required: true
        schema:
          $ref: '#/definitions/dentist.CreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dentist.dentistResponse'
        "400":
          description: validation error
          schema:
            $ref: '#/definitions/problem.Details'
        "404":
          description: clinic not found
          schema:
            $ref: '#/definitions/problem.Details'
        "409":
          description: email already belongs to another dentist in this clinic
          schema:
            $ref: '#/definitions/problem.Details'
      summary: Create a dentist in a clinic
      tags:
      - dentists
  /clinics/{clinic_id}/dentists/{id}:
    delete:
      description: Rejected with 409 if the dentist is the clinic's last administrator
        or last legal representative.
      parameters:
      - description: Clinic ID
        format: uuid
        in: path
        name: clinic_id
        required: true
        type: string
      - description: Dentist ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: dentist deleted
        "400":
          description: id is not a valid UUID
          schema:
            $ref: '#/definitions/problem.Details'
        "404":
          description: clinic or dentist not found
          schema:
            $ref: '#/definitions/problem.Details'
        "409":
          description: would leave the clinic without an administrator or legal representative
          schema:
            $ref: '#/definitions/problem.Details'
      summary: Soft-delete a dentist
      tags:
      - dentists
    get:
      parameters:
      - description: Clinic ID
        format: uuid
        in: path
        name: clinic_id
        required: true
        type: string
      - description: Dentist ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dentist.dentistResponse'
        "400":
          description: id is not a valid UUID
          schema:
            $ref: '#/definitions/problem.Details'
        "404":
          description: clinic or dentist not found
          schema:
            $ref: '#/definitions/problem.Details'
      summary: Get a dentist by id
      tags:
      - dentists
    put:
      consumes:
      - application/json
      description: Only non-null fields in the body are applied. Does not change is_administrator/is_legal_representative
        — use PATCH .../roles for that.
      parameters:
      - description: Clinic ID
        format: uuid
        in: path
        name: clinic_id
        required: true
        type: string
      - description: Dentist ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: dentist
        required: true
        schema:
          $ref: '#/definitions/dentist.UpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dentist.dentistResponse'
        "400":
          description: validation error
          schema:
            $ref: '#/definitions/problem.Details'
        "404":
          description: clinic or dentist not found
          schema:
            $ref: '#/definitions/problem.Details'
        "409":
          description: email already belongs to another dentist in this clinic
          schema:
            $ref: '#/definitions/problem.Details'
      summary: Update a dentist (partial)
      tags:
      - dentists
  /clinics/{clinic_id}/dentists/{id}/roles:
    patch:
      consumes:
      - application/json
      description: An active clinic must always keep at least one administrator and
        one legal representative; demoting the last one of either role returns 409.
      parameters:
      - description: Clinic ID
        format: uuid
        in: path
        name: clinic_id
        required: true
        type: string
      - description: Dentist ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Flags to update (at least one required)
        in: body
        name: roles
        required: true
        schema:
          $ref: '#/definitions/dentist.RolesInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dentist.dentistResponse'
        "400":
          description: validation error
          schema:
            $ref: '#/definitions/problem.Details'
        "404":
          description: clinic or dentist not found
          schema:
            $ref: '#/definitions/problem.Details'
        "409":
          description: would leave the clinic without an administrator or legal representative
          schema:
            $ref: '#/definitions/problem.Details'
      summary: Update a dentist's administrator/legal representative flags
      tags:
      - dentists
...
schemes:
- http
swagger: "2.0"