# clinic-api/internal/dentist/handler.go
package dentist

import (
	"encoding/json"
	"errors"
	"log/slog"
	"net/http"
	"strconv"

	"github.com/go-chi/chi/v5"
	"github.com/google/uuid"

	"github.com/josuesantos1/desafio/pkg/problem"
)

func RegisterRoutes(r chi.Router, svc *Service) {
	h := &handler{svc: svc}
	r.Post("/clinics/{clinic_id}/dentists", h.create)
	r.Get("/clinics/{clinic_id}/dentists", h.list)
	r.Get("/clinics/{clinic_id}/dentists/{id}", h.get)
	r.Put("/clinics/{clinic_id}/dentists/{id}", h.update)
	r.Patch("/clinics/{clinic_id}/dentists/{id}/roles", h.updateRoles)
	r.Delete("/clinics/{clinic_id}/dentists/{id}", h.delete)
}

type handler struct {
	svc *Service
}

// @Summary      Create a dentist in a clinic
// @Description  Creates a dentist attached to the given clinic. Set is_administrator/is_legal_representative afterwards via the roles endpoint if needed.
// @Tags         dentists
// @Accept       json
// @Produce      json
// @Param        clinic_id  path      string       true  "Clinic ID"  format(uuid)
// @Param        dentist    body      CreateInput  true  "Dentist to create"
// @Success      201        {object}  dentistResponse
// @Failure      400        {object}  problem.Details  "validation error"
// @Failure      404        {object}  problem.Details  "clinic not found"
// @Failure      409        {object}  problem.Details  "email already belongs to another dentist in this clinic"
// @Router       /clinics/{clinic_id}/dentists [post]
func (h *handler) create(w http.ResponseWriter, r *http.Request) {
	clinicID, ok := parseUUID(w, chi.URLParam(r, "clinic_id"))
	if !ok {
		return
	}

	var in CreateInput
	if err := json.NewDecoder(r.Body).Decode(&in); err != nil {
		problem.Write(w, http.StatusBadRequest, "VALIDATION_ERROR", "invalid request body", nil)
		return
	}

	d, err := h.svc.Create(r.Context(), clinicID, in)
	if err != nil {
		writeDomainError(w, err)
		return
	}

	problem.WriteJSON(w, http.StatusCreated, toDentistResponse(d))
}

// @Summary      Get a dentist by id
// @Tags         dentists
// @Produce      json
// @Param        clinic_id  path      string  true  "Clinic ID"  format(uuid)
// @Param        id         path      string  true  "Dentist ID"  format(uuid)
// @Success      200        {object}  dentistResponse
// @Failure      400        {object}  problem.Details  "id is not a valid UUID"
// @Failure      404        {object}  problem.Details  "clinic or dentist not found"
// @Router       /clinics/{clinic_id}/dentists/{id} [get]
func (h *handler) get(w http.ResponseWriter, r *http.Request) {
	clinicID, id, ok := parsePathIDs(w, r)
	if !ok {
		return
	}

	d, err := h.svc.Get(r.Context(), clinicID, id)
	if err != nil {
		writeDomainError(w, err)
		return
	}

	problem.WriteJSON(w, http.StatusOK, toDentistResponse(d))
}

// @Summary      Update a dentist (partial)
// @Description  Only non-null fields in the body are applied. Does not change is_administrator/is_legal_representative — use PATCH .../roles for that.
// @Tags         dentists
// @Accept       json
// @Produce      json
// @Param        clinic_id  path      string       true  "Clinic ID"  format(uuid)
// @Param        id         path      string       true  "Dentist ID"  format(uuid)
// @Param        dentist    body      UpdateInput  true  "Fields to update"
// @Success      200        {object}  dentistResponse
// @Failure      400        {object}  problem.Details  "validation error"
// @Failure      404        {object}  problem.Details  "clinic or dentist not found"
// @Failure      409        {object}  problem.Details  "email already belongs to another dentist in this clinic"
// @Router       /clinics/{clinic_id}/dentists/{id} [put]
func (h *handler) update(w http.ResponseWriter, r *http.Request) {
	clinicID, id, ok := parsePathIDs(w, r)
	if !ok {
		return
	}

	var in UpdateInput
	if err := json.NewDecoder(r.Body).Decode(&in); err != nil {
		problem.Write(w, http.StatusBadRequest, "VALIDATION_ERROR", "invalid request body", nil)
		return
	}

	d, err := h.svc.Update(r.Context(), clinicID, id, in)
	if err != nil {
		writeDomainError(w, err)
		return
	}

	problem.WriteJSON(w, http.StatusOK, toDentistResponse(d))
}

// @Summary      Update a dentist's administrator/legal representative flags
// @Description  An active clinic must always keep at least one administrator and one legal representative; demoting the last one of either role returns 409.
// @Tags         dentists
// @Accept       json
// @Produce      json
// @Param        clinic_id  path      string      true  "Clinic ID"  format(uuid)
// @Param        id         path      string      true  "Dentist ID"  format(uuid)
// @Param        roles      body      RolesInput  true  "Flags to update (at least one required)"
// @Success      200        {object}  dentistResponse
// @Failure      400        {object}  problem.Details  "validation error"
// @Failure      404        {object}  problem.Details  "clinic or dentist not found"
// @Failure      409        {object}  problem.Details  "would leave the clinic without an administrator or legal representative"
// @Router       /clinics/{clinic_id}/dentists/{id}/roles [patch]
func (h *handler) updateRoles(w http.ResponseWriter, r *http.Request) {
	clinicID, id, ok := parsePathIDs(w, r)
	if !ok {
		return
	}

	var in RolesInput
	if err := json.NewDecoder(r.Body).Decode(&in); err != nil {
		problem.Write(w, http.StatusBadRequest, "VALIDATION_ERROR", "invalid request body", nil)
		return
	}

	d, err := h.svc.UpdateRoles(r.Context(), clinicID, id, in)
	if err != nil {
		writeDomainError(w, err)
		return
	}

	problem.WriteJSON(w, http.StatusOK, toDentistResponse(d))
}

// @Summary      Soft-delete a dentist
// @Description  Rejected with 409 if the dentist is the clinic's last administrator or last legal representative.
// @Tags         dentists
// @Param        clinic_id  path  string  true  "Clinic ID"  format(uuid)
// @Param        id         path  string  true  "Dentist ID"  format(uuid)
// @Success      204  "dentist deleted"
// @Failure      400  {object}  problem.Details  "id is not a valid UUID"
// @Failure      404  {object}  problem.Details  "clinic or dentist not found"
// @Failure      409  {object}  problem.Details  "would leave the clinic without an administrator or legal representative"
// @Router       /clinics/{clinic_id}/dentists/{id} [delete]
func (h *handler) delete(w http.ResponseWriter, r *http.Request) {
	clinicID, id, ok := parsePathIDs(w, r)
	if !ok {
		return
	}

	if err := h.svc.Delete(r.Context(), clinicID, id); err != nil {
		writeDomainError(w, err)
		return
	}

	w.WriteHeader(http.StatusNoContent)
}

// @Summary      List dentists of a clinic (paginated)
// @Tags         dentists
// @Produce      json
// @Param        clinic_id  path      string  true   "Clinic ID"  format(uuid)
// @Param        limit      query     int     false  "Page size"  default(20)  maximum(100)
// @Param        offset     query     int     false  "Offset for pagination"  default(0)
// @Param        is_administrator        query  bool  false  "Filter by administrator flag"
// @Param        is_legal_representative query  bool  false  "Filter by legal representative flag"
// @Success      200        {object}  listResponse
// @Failure      400        {object}  problem.Details  "clinic_id is not a valid UUID"
// @Failure      404        {object}  problem.Details  "clinic not found"
// @Router       /clinics/{clinic_id}/dentists [get]
func (h *handler) list(w http.ResponseWriter, r *http.Request) {
	clinicID, ok := parseUUID(w, chi.URLParam(r, "clinic_id"))
	if !ok {
		return
	}

	params := ListParams{
		Limit:                 clampLimit(parseQueryInt(r, "limit")),
		Offset:                clampOffset(parseQueryInt(r, "offset")),
		IsAdministrator:       parseQueryBoolPtr(r, "is_administrator"),
		IsLegalRepresentative: parseQueryBoolPtr(r, "is_legal_representative"),
	}

	result, err := h.svc.List(r.Context(), clinicID, params)
	if err != nil {
		writeDomainError(w, err)
		return
	}

	problem.WriteJSON(w, http.StatusOK, toListResponse(result, params))
}

func parseQueryInt(r *http.Request, key string) int {
	v, err := strconv.Atoi(r.URL.Query().Get(key))
	if err != nil {
		return 0
	}
	return v
}

func parseQueryBoolPtr(r *http.Request, key string) *bool {
	raw := r.URL.Query().Get(key)
	if raw == "" {
		return nil
	}
	v, err := strconv.ParseBool(raw)
	if err != nil {
		return nil
	}
	return &v
}

func parsePathIDs(w http.ResponseWriter, r *http.Request) (clinicID, id string, ok bool) {
	clinicID, ok = parseUUID(w, chi.URLParam(r, "clinic_id"))
	if !ok {
		return "", "", false
	}
	id, ok = parseUUID(w, chi.URLParam(r, "id"))
	if !ok {
		return "", "", false
	}
	return clinicID, id, true
}

func parseUUID(w http.ResponseWriter, raw string) (string, bool) {
	if _, err := uuid.Parse(raw); err != nil {
		problem.Write(w, http.StatusBadRequest, "INVALID_ID", "id must be a valid UUID", nil)
		return "", false
	}
	return raw, true
}

func writeDomainError(w http.ResponseWriter, err error) {
	var validationErr *ValidationError
	switch {
	case errors.As(err, &validationErr):
		problem.Write(w, http.StatusBadRequest, "VALIDATION_ERROR", "request validation failed", validationErr.Fields)
	case errors.Is(err, ErrClinicNotFound):
		problem.Write(w, http.StatusNotFound, "CLINIC_NOT_FOUND", "clinic not found", nil)
	case errors.Is(err, ErrEmailExists):
		problem.Write(w, http.StatusConflict, "EMAIL_ALREADY_EXISTS", "email already belongs to another dentist in this clinic", nil)
	case errors.Is(err, ErrNotFound):
		problem.Write(w, http.StatusNotFound, "DENTIST_NOT_FOUND", "dentist not found", nil)
	case errors.Is(err, ErrLastAdminRequired):
		problem.Write(w, http.StatusConflict, "LAST_ADMIN_REQUIRED", "clinic must keep at least one active administrator", nil)
	case errors.Is(err, ErrLastLegalRepresentativeRequired):
		problem.Write(w, http.StatusConflict, "LAST_LEGAL_REPRESENTATIVE_REQUIRED", "clinic must keep at least one active legal representative", nil)
	default:
		slog.Error("unclassified dentist repository error", "error", err)
		problem.Write(w, http.StatusInternalServerError, "INTERNAL_ERROR", "internal server error", nil)
	}
}